tests: add tests for stop_live
